ai-server/requirements.txt
numpy==1.26.4
fastapi
uvicorn[standard]
transformers
timm
requests
Pillow
pydantic-settings>=2.0.0
aio-pika>=9.0.0
asyncpg>=0.29.0
